client::project: renamed project list
